requirements.txt
Django==1.9
django-extensions==1.6.1
ipython==4.0.1
djangorestframework==3.3.2
